Bug behoben und Tests durchgeführt. Läuft. This closes #18.

+++ rest/src/main/java/windyoak/rest/impl/TagsResourceImpl.java
@@ -14,6 +14,7 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriInfo;
 import windyoak.core.OakCoreException;
+import windyoak.core.Project;
 import windyoak.core.StoreService;
 import windyoak.core.Tag;
 import windyoak.core.Tags;
@@ -40,22 +41,22 @@ public class TagsResourceImpl implements TagsResource {
     public Response createTag(UriInfo uriInfo, String tagName, String description) {
         Tag tag = new Tag();
         try {
-            
+
             if (tagName == null || tagName.isEmpty() || description == null
                     || description.isEmpty()) {
                 return Response.status(Response.Status.NOT_ACCEPTABLE).entity("Empty Parameter").build();
             }
             Pattern tp = Pattern.compile("^\\w+");
             Matcher tm = tp.matcher(tagName);
-            if (!tm.matches()){
-              return Response.status(Response.Status.NOT_ACCEPTABLE).entity("Only whole words possible without special characters!").build();
+            if (!tm.matches()) {
+                return Response.status(Response.Status.NOT_ACCEPTABLE).entity("Only whole words possible without special characters!").build();
             }
             if (storeService.getTagByName(tagName) == null) {
                 tag.setName(tagName);
                 tag.setDescription(description);
                 storeService.createTag(tag);
-            }else{
-            return Response.status(Response.Status.NOT_ACCEPTABLE).entity("Tag "+tagName+" already exists!").build();
+            } else {
+                return Response.status(Response.Status.NOT_ACCEPTABLE).entity("Tag " + tagName + " already exists!").build();
             }
 
         } catch (OakCoreException ex) {
@@ -102,13 +103,13 @@ public class TagsResourceImpl implements TagsResource {
             if (tagName == null || tagName.isEmpty()) {
                 return Response.status(Response.Status.NOT_ACCEPTABLE).entity("Empty Parameter!").build();
             }
-            
+
             Pattern tp = Pattern.compile("^\\w+");
             Matcher tm = tp.matcher(tagName);
-            if (!tm.matches()){
-              return Response.status(Response.Status.NOT_ACCEPTABLE).entity("Only whole words possible without special characters!").build();
+            if (!tm.matches()) {
+                return Response.status(Response.Status.NOT_ACCEPTABLE).entity("Only whole words possible without special characters!").build();
             }
-            
+
             newTag = storeService.getTagByName(tagName);
             if (newTag == null) {
                 return Response.status(Response.Status.NOT_FOUND).entity("Can't found Tagname!").build();
@@ -128,9 +129,15 @@ public class TagsResourceImpl implements TagsResource {
             return Response.status(Response.Status.NOT_ACCEPTABLE).entity("Empty Paramter!").build();
         }
         try {
-            Tag newTag = storeService.deleteTag(tagName);
-            //Es muss geprüft werden ob Tag von einem project verwendet wird.
-            return Response.status(Response.Status.OK).entity(newTag).build();
+
+            List<Project> list = storeService.searchProjectByTag(tagName, false);
+            
+            if (list.isEmpty()) {
+                Tag newTag = storeService.deleteTag(tagName);
+                return Response.status(Response.Status.OK).entity(newTag).build();
+            } else {
+                return Response.status(Response.Status.CONFLICT).entity("Tag is used in a Project. Delete the Project first!").build();
+            }
 
         } catch (OakCoreException ex) {
             return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(ex.getMessage()).build();
